Replaced JSON with DB - Can now send workouts from UI directly to DB, and read from DB to graph workouts.

+++ WeightTracker/Exercise.cs
@@ -53,7 +53,7 @@ namespace WeightTracker
         }
 
         // gets the average weight lifted for each rep
-        public int AverageWeightLifted()
+        public float AverageWeightLifted()
         {
             return this.TotalWeightLifted() / this.TotalReps();
         }
